seperate general functions to utils
