+++ scraper/douban_comments_count.py
@@ -4,6 +4,7 @@ from bs4 import BeautifulSoup
 from config import COUNT_TABLE_NAME, BASE_URL, HEADERS
 from db import get_db_conn
 import re
+from utils import extract_count, safe_number, safe_float_percent
 
 
 def extract_movie_stats(drama_url, headers=None):
@@ -14,34 +15,7 @@ def extract_movie_stats(drama_url, headers=None):
     response = requests.get(drama_url, headers=headers, timeout=10)
     soup = BeautifulSoup(response.text, 'html.parser')
 
-    def extract_count(soup, pattern, selector=None):
-        try:
-            if selector:
-                # If the selector is for comments, handle it separately 
-                if selector == 'a[href*="comments?status=P"]':
-                    tag = soup.find('a', href=re.compile(r'comments\?status=[PF]'))
-                else:
-                    tag = soup.select_one(selector)
-                # If the tag is not found, return None
-                if tag is None:
-                    print(f"⚠️ Selector not found: {selector}")
-                    return None
-                # Extract text from the tag 
-                text = tag.get_text(strip=True)
-            else:
-                text = soup.get_text()
-            # Use regex to find the count in the text 
-            match = re.search(pattern, text)
-            if match:
-                raw_value = match.group(1).replace(',', '').replace(' ', '')
-                return safe_number(raw_value) 
-            else:
-                print(f"⚠️ Pattern not matched: {pattern}")
-                return None
-            
-        except (AttributeError, TypeError, re.error) as e:
-            print(f"❌ Error extracting count: {e}")
-            return None
+    
 
     rating = extract_count(soup, r'(\d+\.\d+)', 'strong[property="v:average"]')
     rating_people = extract_count(soup, r'(\d+)', 'span[property="v:votes"]')
@@ -65,35 +39,6 @@ def extract_movie_stats(drama_url, headers=None):
     }
 
 
-def safe_number(value):
-    """
-    Convert value to int if it's integer.
-    Otherwise convert to float rounded to 1 decimal place.
-    Return None if conversion fails.
-    """
-    if value is None:
-        return None
-    try:
-        num = float(value)
-        if num.is_integer():
-            return int(num)
-        else:
-            return round(num, 1)
-    except (ValueError, TypeError):
-        return None
-
-
-def safe_float_percent(value):
-    """ Safely converts a percentage string to a float.
-    :param value: The percentage string to convert
-    :return: float if conversion is successful, None otherwise
-    """ 
-    try:
-        if not value or '%' not in value:
-            return None
-        return float(value.strip('%'))
-    except ValueError:
-        return None
 
 
 def insert_movie_stats(movie_stats, db_conn):
